+++ packages/web/src/app/[lng]/(main)/profile/setting/delete/components/DeleteModal.client.tsx
@@ -1,10 +1,11 @@
 import DeleteCompleteModal from './DeleteCompleteModal.client';
+
 import { usePatchSignOut } from '@/apis/profile';
 import { useTranslation } from '@/app/i18n/client';
 import { Icon } from '@/components/Icon';
 import { Modal } from '@/components/Modal';
 import { Spacing } from '@/components/Spacing';
-import useLogout from '@/hooks/token/useLogout';
+import useLogout from '@/hooks/useLogout';
 import { useModal } from '@/hooks/useModal';
 
 interface DeleteModalProps {
