--- packages/web/src/hooks/common/useDidMount/index.test.ts
@@ -1,7 +1,0 @@
-import useDidMount from './useDidMount';
-
-describe('useDidMount', () => {
-  it('should defined with default', () => {
-    expect(useDidMount).toBeDefined();
-  });
-});
