--- packages/web/src/hooks/common/useDidMount/index.ts
@@ -1,1 +1,0 @@
-export { default as useDidMount } from './useDidMount';
