--- packages/web/src/hooks/common/useDidUnMount/index.ts
@@ -1,1 +1,0 @@
-export { default as useDidUnMount } from './useDidUnMount';
