--- packages/web/src/hooks/common/useDidUpdate/index.ts
@@ -1,1 +1,0 @@
-export { default as useDidUpdate } from './useDidUpdate';
