+++ packages/web/src/hooks/useAppRouter.ts
@@ -2,7 +2,7 @@ import { useRouter } from 'next/navigation';
 
 import { sendMessageToReactNative } from '@/utils';
 
-const useAppRouter = () => {
+export default function useAppRouter() {
   const isApp = false;
   const router = useRouter();
 
@@ -56,6 +56,4 @@ const useAppRouter = () => {
   };
 
   return { push, back, replace, refresh, reset };
-};
-
-export default useAppRouter;
+}
