+++ packages/web/src/hooks/useBrowser.ts
@@ -1,5 +1,6 @@
-import { useDidMount } from './common/useDidMount';
 import { useState } from 'react';
+
+import { useDidMount } from './common/useDidMount';
 
 export default function useBrowser() {
   const [browser, setBrowser] = useState('');
