+++ packages/web/src/hooks/useDidMount.test.tsx
@@ -1,7 +1,8 @@
-import useDidMount from './useDidMount';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { renderHook } from '@testing-library/react-hooks';
 import { useState } from 'react';
+
+import useDidMount from '../../common/useDidMount/useDidMount';
 
 describe('useDidMount', () => {
   it('default export이여야 한다', () => {
