+++ packages/web/src/hooks/useDidUnMount.ts
@@ -1,11 +1,9 @@
 import { useEffect } from 'react';
 
-const useDidUnMount = (callback: VoidFunction) => {
+export default function useDidUnMount(callback: VoidFunction) {
   useEffect(() => {
     return () => {
       callback();
     };
   }, []);
-};
-
-export default useDidUnMount;
+}
