+++ packages/web/src/hooks/useFileUpload.tsx
@@ -1,6 +1,8 @@
-import { getBase64FromImage } from './getBase64FromImage';
-import { usePostFiles } from '@/apis/common';
 import { useState } from 'react';
+
+import getBase64FromImage from './getBase64FromImage';
+
+import { usePostFiles } from '@/apis/common';
 
 interface UseImageUploadProps {
   /**
@@ -19,7 +21,7 @@ interface UseImageUploadProps {
   };
 }
 
-export function useFileUpload(
+export default function useFileUpload(
   handleFileChange: UseImageUploadProps['handleFileChange'],
   options?: UseImageUploadProps['options']
 ) {
