+++ packages/web/src/hooks/useIntersect.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 
-const useIntersect = (callback: () => void, options?: IntersectionObserverInit) => {
+export default function useIntersect(callback: () => void, options?: IntersectionObserverInit) {
   const target = useRef(null);
 
   useEffect(() => {
@@ -11,6 +11,4 @@ const useIntersect = (callback: () => void, options?: IntersectionObserverInit) 
   }, [callback, options]);
 
   return target;
-};
-
-export default useIntersect;
+}
