+++ packages/web/src/hooks/useNumberParams.ts
@@ -2,7 +2,7 @@ import { useParams } from 'next/navigation';
 
 type NonEmptyArray<T> = [T, ...T[]];
 
-export function useNumberParams<T extends NonEmptyArray<string>>() {
+export default function useNumberParams<T extends NonEmptyArray<string>>() {
   const params = useParams();
 
   const numberParams = Object.fromEntries(
