+++ packages/web/src/hooks/useOnClickOutside.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 
-export function useOnClickOutside<T extends HTMLElement = HTMLElement>(
+export default function useOnClickOutside<T extends HTMLElement = HTMLElement>(
   ref: React.RefObject<T>,
   handler: (event: MouseEvent | TouchEvent) => void
 ) {
