+++ packages/web/src/hooks/useShowMore.ts
@@ -1,11 +1,12 @@
-import { useDidMount } from './common/useDidMount';
 import { useEffect, useRef, useState } from 'react';
+
+import { useDidMount } from './common/useDidMount';
 
 interface UseShowMoreProps {
   maxLines: number;
 }
 
-export function useShowMore({ maxLines }: UseShowMoreProps) {
+export default function useShowMore({ maxLines }: UseShowMoreProps) {
   const [showFullText, setShowFullText] = useState(true);
   const [shouldShowButton, setShouldShowButton] = useState(false);
   const contentRef = useRef<HTMLDivElement>(null);
